+++ sample/src/main/java/com/prolificinteractive/materialcalendarview/sample/decorators/HighlightWeekendsDecorator.java
@@ -1,0 +1,48 @@
+package com.prolificinteractive.materialcalendarview.sample.decorators;
+
+import android.graphics.Color;
+import android.graphics.LinearGradient;
+import android.graphics.Shader;
+import android.graphics.drawable.Drawable;
+import android.graphics.drawable.ShapeDrawable;
+import android.graphics.drawable.shapes.RoundRectShape;
+
+import com.prolificinteractive.materialcalendarview.CalendarDay;
+import com.prolificinteractive.materialcalendarview.DayViewDecorator;
+import com.prolificinteractive.materialcalendarview.DayViewFacade;
+
+import java.util.Calendar;
+
+/**
+ * Highlight Saturdays and Sundays with a background
+ */
+public class HighlightWeekendsDecorator implements DayViewDecorator {
+
+    @Override
+    public boolean shouldDecorate(CalendarDay day) {
+        int weekDay = day.getCalendar().get(Calendar.DAY_OF_WEEK);
+        return weekDay == Calendar.SATURDAY || weekDay == Calendar.SUNDAY;
+    }
+
+    @Override
+    public void decorate(DayViewFacade view) {
+        view.setBackgroundUnselected(generateBackgroundDrawable());
+    }
+
+    private static Drawable generateBackgroundDrawable() {
+        final int r = 0;
+        final float[] outerR = new float[] {r, r, r, r, r, r, r, r};
+        final int color = Color.parseColor("#228BC34A");
+
+        RoundRectShape rr = new RoundRectShape(outerR, null, null);
+
+        ShapeDrawable drawable = new ShapeDrawable(rr);
+        drawable.setShaderFactory(new ShapeDrawable.ShaderFactory() {
+            @Override
+            public Shader resize(int width, int height) {
+                return new LinearGradient(0, 0, 0, 0, color, color, Shader.TileMode.REPEAT);
+            }
+        });
+        return drawable;
+    }
+}